Exercise 6.7 solution
